requirements.txt
Django==4.2.4
environs==9.5.0
flake8==6.0.0
pre-commit==3.3.3
dj-database-url==2.1.0
psycopg2-binary==2.9.7
Pillow==10.0.0
django-phonenumber-field[phonenumbers]==7.1.0
